Solution to the issue #2741
Issue: Add filter to choose the amount of categories loaded in category spreading chart.

Added new entry for number of categories loaded in spreading as well as two buttons for incrementing and decrementing number of categories.

Entry is cleared for edge cases as well as any non numeric input.

Number of categories is persisted via SettingsFacade.cs just like in AccountGroup.cs, but I am not so sure with my solution so maybe it needs a correction.

Filter was tested on both Android and Windows and works without problems. Also passed all the existing tests.
However, I did not create any new tests for my changes.

Also text in the new label for the entry is hardcoded in English, I did not add all the necessary translations which I apologize for.